refactor(ui): move handle_app_action into app_component/actions.rs
Last of the app_component split. handle_app_action is where an action lands
once the component hierarchy has declined it, and at 318 lines it was the only
thing keeping mod.rs above 800.

mod.rs is now 475 and reads as what it is: construction, sync scheduling, the
Component impl, handle_event and render.

Pure move, the function body is byte for byte what it was. Two doc comments
stranded by earlier moves go with it: one for handle_global_key left behind by
the keys.rs split, and one for the old spawn_task_operation left above spawn()
by the Operation rework.
